Add session rewriting to Application Module
