+++ dd-java-agent/agent-iast/src/main/java/com/datadog/iast/IastSystem.java
@@ -21,7 +21,6 @@ import com.datadog.iast.sink.NoHttpOnlyCookieModuleImpl;
 import com.datadog.iast.sink.NoSameSiteCookieModuleImpl;
 import com.datadog.iast.sink.PathTraversalModuleImpl;
 import com.datadog.iast.sink.ReflectionInjectionModuleImpl;
-import com.datadog.iast.sink.SessionRewritingModuleImpl;
 import com.datadog.iast.sink.SqlInjectionModuleImpl;
 import com.datadog.iast.sink.SsrfModuleImpl;
 import com.datadog.iast.sink.StacktraceLeakModuleImpl;
@@ -148,8 +147,7 @@ public class IastSystem {
             ApplicationModuleImpl.class,
             HardcodedSecretModuleImpl.class,
             InsecureAuthProtocolModuleImpl.class,
-            ReflectionInjectionModuleImpl.class,
-            SessionRewritingModuleImpl.class);
+            ReflectionInjectionModuleImpl.class);
     if (iast != FULLY_ENABLED) {
       modules = modules.filter(IastSystem::isOptOut);
     }
