+++ dd-java-agent/instrumentation/servlet/request-3/src/main/java/datadog/trace/instrumentation/servlet3/IastServlet3Advice.java
@@ -4,7 +4,6 @@ import datadog.trace.api.iast.InstrumentationBridge;
 import datadog.trace.api.iast.Sink;
 import datadog.trace.api.iast.VulnerabilityTypes;
 import datadog.trace.api.iast.sink.ApplicationModule;
-import datadog.trace.api.iast.sink.SessionRewritingModule;
 import datadog.trace.bootstrap.InstrumentationContext;
 import java.util.HashSet;
 import java.util.Set;
@@ -20,8 +19,7 @@ public class IastServlet3Advice {
   @Advice.OnMethodExit(suppress = Throwable.class)
   public static void onExit(@Advice.Argument(0) ServletRequest request) {
     final ApplicationModule applicationModule = InstrumentationBridge.APPLICATION;
-    final SessionRewritingModule sessionRewritingModule = InstrumentationBridge.SESSION_REWRITING;
-    if (applicationModule == null && sessionRewritingModule == null) {
+    if (applicationModule == null) {
       return;
     }
     if (!(request instanceof HttpServletRequest)) {
@@ -34,15 +32,14 @@ public class IastServlet3Advice {
     InstrumentationContext.get(ServletContext.class, Boolean.class).put(context, true);
     if (applicationModule != null) {
       applicationModule.onRealPath(context.getRealPath("/"));
-    }
-    if (sessionRewritingModule != null
-        && context.getEffectiveSessionTrackingModes() != null
-        && !context.getEffectiveSessionTrackingModes().isEmpty()) {
-      Set<String> sessionTrackingModes = new HashSet<>();
-      for (SessionTrackingMode mode : context.getEffectiveSessionTrackingModes()) {
-        sessionTrackingModes.add(mode.name());
+      if (context.getEffectiveSessionTrackingModes() != null
+          && !context.getEffectiveSessionTrackingModes().isEmpty()) {
+        Set<String> sessionTrackingModes = new HashSet<>();
+        for (SessionTrackingMode mode : context.getEffectiveSessionTrackingModes()) {
+          sessionTrackingModes.add(mode.name());
+        }
+        applicationModule.checkSessionTrackingModes(sessionTrackingModes);
       }
-      sessionRewritingModule.checkSessionTrackingModes(sessionTrackingModes);
     }
   }
 }
