+++ dd-java-agent/instrumentation/servlet/request-3/src/test/groovy/JettyServlet3Test.groovy
@@ -1,7 +1,6 @@
 import datadog.trace.agent.test.asserts.TraceAssert
 import datadog.trace.agent.test.base.HttpServer
 import datadog.trace.agent.test.naming.TestingGenericHttpNamingConventions
-import datadog.trace.api.iast.sink.SessionRewritingModule
 import datadog.trace.api.iast.InstrumentationBridge
 import datadog.trace.api.iast.sink.ApplicationModule
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan
@@ -533,7 +532,6 @@ class IastJettyServlet3ForkedTest extends JettyServlet3TestSync {
   void 'test no calls if no modules registered'() {
     given:
     final appModule = Mock(ApplicationModule)
-    final sessionRewritingModule = Mock(SessionRewritingModule)
     def request = request(SUCCESS, "GET", null).build()
 
     when:
@@ -541,16 +539,14 @@ class IastJettyServlet3ForkedTest extends JettyServlet3TestSync {
 
     then:
     0 * appModule.onRealPath(_)
-    0 * sessionRewritingModule.checkSessionTrackingModes(_)
+    0 * appModule.checkSessionTrackingModes(_)
     0 * _
   }
 
   void 'test that iast modules are called'() {
     given:
     final appModule = Mock(ApplicationModule)
-    final sessionRewritingModule = Mock(SessionRewritingModule)
     InstrumentationBridge.registerIastModule(appModule)
-    InstrumentationBridge.registerIastModule(sessionRewritingModule)
     def request = request(SUCCESS, "GET", null).build()
 
     when:
@@ -558,7 +554,7 @@ class IastJettyServlet3ForkedTest extends JettyServlet3TestSync {
 
     then:
     1 *  appModule.onRealPath(_)
-    1 *  sessionRewritingModule.checkSessionTrackingModes(_)
+    1 *  appModule.checkSessionTrackingModes(_)
     0 * _
 
     when:
@@ -566,7 +562,7 @@ class IastJettyServlet3ForkedTest extends JettyServlet3TestSync {
 
     then: //Only call once per application context
     0 *  appModule.onRealPath(_)
-    0 *  sessionRewritingModule.checkSessionTrackingModes(_)
+    0 *  appModule.checkSessionTrackingModes(_)
     0 * _
   }
