+++ dd-java-agent/instrumentation/servlet/request-5/src/main/java/datadog/trace/instrumentation/servlet5/IastJakartaServletInstrumentation.java
@@ -12,7 +12,6 @@ import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.api.iast.InstrumentationBridge;
 import datadog.trace.api.iast.sink.ApplicationModule;
-import datadog.trace.api.iast.sink.SessionRewritingModule;
 import datadog.trace.bootstrap.InstrumentationContext;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.SessionTrackingMode;
@@ -64,8 +63,7 @@ public class IastJakartaServletInstrumentation extends InstrumenterModule.Iast
     @Advice.OnMethodExit(suppress = Throwable.class)
     public static void after(@Advice.This final HttpServlet servlet) {
       final ApplicationModule applicationModule = InstrumentationBridge.APPLICATION;
-      final SessionRewritingModule sessionRewritingModule = InstrumentationBridge.SESSION_REWRITING;
-      if (applicationModule == null && sessionRewritingModule == null) {
+      if (applicationModule == null) {
         return;
       }
       final ServletContext context = servlet.getServletContext();
@@ -75,15 +73,14 @@ public class IastJakartaServletInstrumentation extends InstrumenterModule.Iast
       InstrumentationContext.get(ServletContext.class, Boolean.class).put(context, true);
       if (applicationModule != null) {
         applicationModule.onRealPath(context.getRealPath("/"));
-      }
-      if (sessionRewritingModule != null
-          && context.getEffectiveSessionTrackingModes() != null
-          && !context.getEffectiveSessionTrackingModes().isEmpty()) {
-        Set<String> sessionTrackingModes = new HashSet<>();
-        for (SessionTrackingMode mode : context.getEffectiveSessionTrackingModes()) {
-          sessionTrackingModes.add(mode.name());
+        if (context.getEffectiveSessionTrackingModes() != null
+            && !context.getEffectiveSessionTrackingModes().isEmpty()) {
+          Set<String> sessionTrackingModes = new HashSet<>();
+          for (SessionTrackingMode mode : context.getEffectiveSessionTrackingModes()) {
+            sessionTrackingModes.add(mode.name());
+          }
+          applicationModule.checkSessionTrackingModes(sessionTrackingModes);
         }
-        sessionRewritingModule.checkSessionTrackingModes(sessionTrackingModes);
       }
     }
   }
